// client/src/components/Dashboard/Models/Venue.ts
export default interface IVenue {
    id: string;
    name: string;
    contact: any;
    location: any;
    categories: any;
    verified: boolean;
    stats: any;
    beenHere: any;
    specials: any;
    venuePage: any;
    hereNow: any;
    referralId: string;
    venueChains: any;
    hasPerk: boolean;
}